cloud_computing_project/counter/serializers.py
from django.db.models import fields
from rest_framework import serializers
from .models import Counter

class CounterSerializer(serializers.ModelSerializer):
    class Meta:
        model = Counter
        fields = ['id', 'value']